@@ -1,14 +1,11 @@
 {
-  "name": "@restqa/ui",
-  "version": "0.1.0",
-  "private": false,
+  "name": "@restga/ui",
+  "version": "1.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "lint:fix": "vue-cli-service lint --fix",
-    "test": "vue-cli-service test:unit",
-    "test:watch": "vue-cli-service test:unit --watch"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
     "@chenfengyuan/vue-countdown": "^2.0.0-beta",
@@ -20,31 +17,25 @@
     "vue": "^3.0.0",
     "vue-gtag-next": "^1.14.0",
     "vue-prism-editor": "^2.0.0-alpha.2",
-    "vue-router": "^4.0.8",
+    "vue-router": "^4.1.5",
     "vue3-clipboard": "^1.0.0",
     "vue3-highcharts": "^0.1.3",
-    "vuex": "^4.0.1",
-    "yaml": "^2.0.1"
+    "vuex": "^4.1.0",
+    "yaml": "^2.1.3"
   },
+  "type": "module",
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "^4.5.13",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/compiler-sfc": "^3.0.11",
-    "@vue/test-utils": "^2.0.0-rc.8",
-    "babel-eslint": "^10.1.0",
-    "copy-webpack-plugin": "^6.4.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0",
-    "identity-obj-proxy": "^3.0.0",
-    "sass": "^1.34.0",
-    "sass-loader": "^10.2.0",
-    "vue-jest": "^5.0.0-alpha.10"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@rushstack/eslint-patch": "^1.1.4",
+    "@vitejs/plugin-vue": "^3.1.2",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "eslint": "^8.22.0",
+    "eslint-plugin-vue": "^9.3.0",
+    "prettier": "^2.7.1",
+    "sass": "^1.56.0",
+    "unplugin-auto-import": "^0.11.4",
+    "unplugin-vue-components": "^0.22.9",
+    "vite": "^3.1.8",
+    "vite-plugin-singlefile": "^0.12.3",
+    "vite-plugin-static-copy": "^0.11.1"
+  }
 }
