@@ -1,6 +1,6 @@
 <template>
   <Card class="card" :title="title">
-    <img :src="require(`@/assets/images/logos/${name}.png`)" />
+    <img :src="`images/logos/${name}.png`" />
     <br />
     <br />
     <el-link v-if="location" :href="location" target="_blank"
@@ -9,29 +9,30 @@
   </Card>
 </template>
 <script>
-import Card from "@/components/UI/card/Card";
+import Card from "@/components/UI/card/Card.vue";
 
 export default {
   name: "CollectionTool",
   props: {
     name: {
       type: String,
-      required: true
+      required: true,
     },
     location: {
       type: String,
-      required: true
-    }
+      required: true,
+    },
   },
   components: {
-    Card
+    Card,
   },
+  method: {},
   computed: {
     title() {
       const name = this.name.charAt(0).toUpperCase() + this.name.slice(1);
       return name + (this.location ? "" : " (coming soon)");
-    }
-  }
+    },
+  },
 };
 </script>
 <style scoped>
