@@ -1,13 +1,6 @@
 <template>
-  <card
-    title="Documentation"
-    emoji="📖"
-    class="card"
-  >
-    <img
-      :src="require(`@/assets/images/mascot/documentation.png`)"
-      class="option-img"
-    />
+  <card title="Documentation" emoji="📖" class="card">
+    <img :src="`images/mascot/${key}.png`" class="option-img" />
     <br /><br />
 
     <el-link type="primary" v-on:click.stop.prevent="goTo()"
@@ -17,24 +10,29 @@
 </template>
 
 <script>
-import Card from "@/components/UI/card/Card";
+import Card from "@/components/UI/card/Card.vue";
 
 export default {
   name: "DocumentationWidget",
   components: {
-    Card
+    Card,
+  },
+  data() {
+    return {
+      key: "documentation",
+    };
   },
   methods: {
     goTo() {
       const route = {
         name: "documentationPage",
         params: {
-          id: "introduction"
-        }
-      }
+          id: "introduction",
+        },
+      };
       this.$router.push(route);
-    }
-  }
+    },
+  },
 };
 </script>
 <style scoped>
@@ -46,5 +44,3 @@ img {
   height: 100px;
 }
 </style>
-
-
