@@ -5,14 +5,14 @@
   </card>
 </template>
 <script>
-import Card from "@/components/UI/card/Card";
+import Card from "@/components/UI/card/Card.vue";
 import VueHighcharts from "vue3-highcharts";
 
 export default {
   name: "RestqaTestChartFeatures",
   components: {
     Card,
-    VueHighcharts
+    VueHighcharts,
   },
   methods: {
     getScenarioData() {
@@ -20,22 +20,22 @@ export default {
       return [
         {
           label: "Passed",
-          data: data.passed
+          data: data.passed,
         },
         {
           label: "Failed",
-          data: data.failed
+          data: data.failed,
         },
         {
           label: "skipped",
-          data: data.skipped
+          data: data.skipped,
         },
         {
           label: "undefined",
-          data: data.undefined
-        }
+          data: data.undefined,
+        },
       ];
-    }
+    },
   },
   computed: {
     scenariosForChart() {
@@ -50,27 +50,27 @@ export default {
         danger: "#FD5F00",
         inverse: "#0D2B47",
         textColor: "#495057",
-        gray: "#D7DFE6"
+        gray: "#D7DFE6",
       };
-      let {success, danger, info, secondary} = colors;
+      let { success, danger, info, secondary } = colors;
       let series = [
         {
           name: "Scenarios",
           data: this.getScenarioData().map((s) => {
             return {
               name: s.label,
-              y: s.data
+              y: s.data,
             };
-          })
-        }
+          }),
+        },
       ];
       return {
         chart: {
           type: "pie",
-          height: 300
+          height: 300,
         },
         credits: {
-          enabled: false
+          enabled: false,
         },
         title: false,
         plotOptions: {
@@ -78,7 +78,7 @@ export default {
             dataLabels: {
               enabled: true,
               format:
-                "<b>{point.name}</b>:<br>{point.percentage:.1f} %<br>value: {point.y}"
+                "<b>{point.name}</b>:<br>{point.percentage:.1f} %<br>value: {point.y}",
             },
             borderColor: null,
             //showInLegend: true,
@@ -87,11 +87,11 @@ export default {
             states: {
               hover: {
                 halo: {
-                  size: 1
-                }
-              }
-            }
-          }
+                  size: 1,
+                },
+              },
+            },
+          },
         },
         colors: [success, danger, info, secondary],
         legend: {
@@ -101,18 +101,18 @@ export default {
           itemStyle: {
             color: "#495057",
             fontWeight: 100,
-            fontFamily: "Montserrat"
+            fontFamily: "Montserrat",
           },
           itemMarginBottom: 5,
-          symbolRadius: 0
+          symbolRadius: 0,
         },
         exporting: {
-          enabled: false
+          enabled: false,
         },
-        series
+        series,
       };
-    }
-  }
+    },
+  },
 };
 </script>
 
