@@ -1,21 +1,12 @@
 <template>
   <div class="menu">
-    <restqa-select-config v-if="hasConfig"></restqa-select-config>
     <el-avatar class="avatar" icon="el-icon-user-solid"></el-avatar>
   </div>
 </template>
 <script>
-import RestqaSelectConfig from "@/components/restqa/select-config/RestQASelectConfig.vue";
-
 export default {
-  components: {
-    RestqaSelectConfig
-  },
-  computed: {
-    hasConfig() {
-      return Boolean(this.$store.getters.config);
-    }
-  }
+  name: "HeaderPartial",
+  components: {},
 };
 </script>
 
