@@ -1,34 +1,3 @@
-const mutations = {
-  info(state, obj) {
-    state.info = obj;
-  },
-  config(state, obj) {
-    state.config = obj;
-  },
-  loadingConfig(state, bool) {
-    state.loadingConfig = bool;
-  },
-  selectedEnv(state, val) {
-    state.selectedEnv = val;
-  },
-  steps(state, val) {
-    state.steps = val;
-  },
-  features(state, val) {
-    state.features = val;
-  },
-  selectedFile(state, val) {
-    state.selectedFile = val;
-  },
-  testResult(state, val) {
-    state.testResult = val;
-  },
-  projectStatus(state, val) {
-    state.projectStatus = val;
-  },
-  preferences(state, preferencesObject) {
-    state.preferences = preferencesObject;
-  }
-};
+const mutations = {};
 
 export default mutations;
