@@ -1,6 +1,8 @@
 <template>
   <el-breadcrumb separator=">">
-    <el-breadcrumb-item :to="{name: 'homepage'}">Dashboard</el-breadcrumb-item>
+    <el-breadcrumb-item :to="{ name: 'homepage' }"
+      >Dashboard</el-breadcrumb-item
+    >
     <el-breadcrumb-item>API Specification</el-breadcrumb-item>
   </el-breadcrumb>
   <h2><i class="el-icon-collection" /> API Specification</h2>
@@ -21,14 +23,14 @@ import Specification from "@/components/restqa/specification/Specification.vue";
 export default {
   components: {
     Editor,
-    Specification
+    Specification,
   },
-  name: "Openapi",
+  name: "OpenapiPage",
   data() {
     const specification = this.$store.getters.projectStatus.specification;
     return {
-      specification
+      specification,
     };
-  }
+  },
 };
 </script>
